moved helpers.GetRandomString to getRandomString
