+++ internal/client/helpers_test.go
@@ -1,10 +1,9 @@
-package helpers
+package client
 
 import "testing"
-
 func TestGetRandomStr(t *testing.T) {
 	wantSize := 32
-	testStr := GetRandomSrt(wantSize)
+	testStr := getRandomSrt(wantSize)
 	haveSize := len(testStr)
 	if wantSize != haveSize {
 		t.Errorf("string lenght mismatch: want: %d, have: %d",
